[refactor] 전체 감상 / 무한스크롤 로딩 아이템 생성 (#458)
