+++ android/app/src/main/java/com/teamtripdraw/android/ui/allPosts/AllPostsViewModel.kt
@@ -2,18 +2,17 @@ package com.teamtripdraw.android.ui.allPosts
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.Transformations
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.teamtripdraw.android.TripDrawApplication
 import com.teamtripdraw.android.domain.model.post.PostOfAll
 import com.teamtripdraw.android.domain.repository.PostRepository
 import com.teamtripdraw.android.support.framework.presentation.event.Event
-import com.teamtripdraw.android.ui.model.UiAllPosts
+import com.teamtripdraw.android.ui.model.allPosts.UiAllPosts
+import com.teamtripdraw.android.ui.model.allPosts.UiLoading
 import com.teamtripdraw.android.ui.model.mapper.toPresentation
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
-import java.time.LocalDateTime
 import javax.inject.Inject
 
 @HiltViewModel
@@ -22,22 +21,13 @@ class AllPostsViewModel @Inject constructor(
 ) : ViewModel() {
 
     private val _posts: MutableLiveData<List<PostOfAll>> = MutableLiveData(listOf())
-    val posts: LiveData<UiAllPosts> =
-        Transformations.map(_posts) { post -> UiAllPosts(post.map { it.toPresentation() }) }
+    private val _uiPosts: MutableLiveData<UiAllPosts> =
+        MutableLiveData(UiAllPosts(_posts.value!!.map { it.toPresentation() }))
+    val posts: LiveData<UiAllPosts> = _uiPosts
+//        Transformations.map(_posts) { post -> UiAllPosts(post.map { it.toPresentation() }) }
 
     private val _openPostDetailEvent = MutableLiveData<Event<Long>>()
     val openPostDetailEvent: LiveData<Event<Long>> = _openPostDetailEvent
-
-    private val loadingItem = PostOfAll(
-        postId = -1,
-        tripId = -1,
-        title = "",
-        writing = "",
-        address = "",
-        postImageUrl = "",
-        routeImageUrl = "",
-        recordedAt = LocalDateTime.MIN,
-    )
 
     private var lastId: Long? = null
 
@@ -56,7 +46,8 @@ class AllPostsViewModel @Inject constructor(
     }
 
     private fun addLoadingItem() {
-        _posts.value = _posts.value!!.toMutableList().apply { add(loadingItem) }
+        _uiPosts.value = _uiPosts.value!!.apply { this.postItems.toMutableList().add(UiLoading()) }
+//        _posts.value = _posts.value!!.toMutableList().apply { add(loadingItem) }
     }
 
     private fun getPosts() {
@@ -83,8 +74,14 @@ class AllPostsViewModel @Inject constructor(
     }
 
     private fun fetchPosts(posts: List<PostOfAll>) {
+//        _posts.value = _posts.value!!.toMutableList().apply {
+//            remove(loadingItem)
+//            addAll(posts)
+//        }
+        _uiPosts.value = _uiPosts.value!!.apply {
+            this.postItems.toMutableList().removeLast()
+        }
         _posts.value = _posts.value!!.toMutableList().apply {
-            remove(loadingItem)
             addAll(posts)
         }
     }
