--- android/app/src/main/java/com/teamtripdraw/android/ui/model/UiAllPosts.kt
@@ -1,6 +1,0 @@
-package com.teamtripdraw.android.ui.model
-
-data class UiAllPosts(
-    val postItems: List<UiPostOfAll>,
-    val isEmpty: Boolean = postItems.isEmpty(),
-)
