+++ android/app/src/main/java/com/teamtripdraw/android/ui/model/allPosts/UiAllPosts.kt
@@ -1,0 +1,6 @@
+package com.teamtripdraw.android.ui.model.allPosts
+
+data class UiAllPosts(
+    val postItems: List<UiItemView>,
+    val isEmpty: Boolean = postItems.isEmpty(),
+)
