+++ android/app/src/main/java/com/teamtripdraw/android/ui/model/allPosts/UiItemView.kt
@@ -1,0 +1,3 @@
+package com.teamtripdraw.android.ui.model.allPosts
+
+interface UiItemView
