+++ android/app/src/main/java/com/teamtripdraw/android/ui/model/allPosts/UiLoading.kt
@@ -1,0 +1,3 @@
+package com.teamtripdraw.android.ui.model.allPosts
+
+class UiLoading : UiItemView
