+++ android/app/src/main/java/com/teamtripdraw/android/ui/model/mapper/PostUiMapper.kt
@@ -5,7 +5,7 @@ import com.teamtripdraw.android.domain.model.post.PostOfAll
 import com.teamtripdraw.android.support.framework.presentation.LocalDateTimeFormatter.displayDateTimeFormatter
 import com.teamtripdraw.android.ui.model.UiPostDetail
 import com.teamtripdraw.android.ui.model.UiPostItem
-import com.teamtripdraw.android.ui.model.UiPostOfAll
+import com.teamtripdraw.android.ui.model.allPosts.UiPostOfAll
 import java.time.LocalDateTime
 
 fun Post.toPresentation() = UiPostItem(
